Test changes

+++ src/main/java/com/majkowski/Product.java
@@ -8,12 +8,12 @@ public class Product {
   /** Nazwa produktu. */
   public final transient String name;
   /** Cena produktu. */
-  public final transient int price;
+  public final transient double price;
   /** Ilosc produktu. */
   public final transient int count;
 
   /** Kostruktor produktu. */
-  public Product(final String name,  final int price, final int count) {
+  public Product(final String name,  final double price, final int count) {
     this.name = name;
     this.price = price;
     this.count = count;
